core: Generalize TrackerDBStatement cache API
This API was so far internal to TrackerDBInterface, in order to cache
statements for selects and updates. Since we want to add other caching
layers elsewhere, make this API public and more generic.

Now, it is possible to define hash/equal/destroy functions, so that it
is not limited to SQL query strings as keys, this will be useful in
future commits. In the mean time, reimplement the select/update caches
on top of this API.