core: Improve performance of update statement caching
Currently the code relies on TrackerDBInterface builtin caching, which
consists on lookups for the SQL string prior to statement compiling/binding.
This is a very late layer of caching, and relies on us creating the full
SQL string (i.e. basically one step away from a sqlite3_stmt) to reuse
a cached statement.

The caching strategy here can be significantly improved, we basically
need a graph|{class|property} tuple to look up a statement, and that is
something that TrackerDataLogEntry structs in the event log already
have. So we may go directly from a log entry to a statement without
intermediate string stages.

And that's precisely what this commit does. There is a new set of
hash/equal functions in order to match these log entries in the
statement MRU, so that lookups are fast based on pointer hashes and
comparisons.

Since we need to preserve a copy of these TrackerDataLogEntry as keys
in the MRU cache, add copy/free functions that handle copying the
necessary data (a copy of the TrackerDataLogEntry, but also partial
copy of the array containing propery changes for this update).

This makes lookups sensibly faster, compared to the old laborious
string building for lookups.