core: Move accounting of already visited TrackerResources a level up
The improvement is twofold, if a same TrackerResource is referenced by
different elements in a TrackerBatch, we will skip second additions
altogether. On the other hand, the hashtable is more long-lived and
not created/freed all the time tens thousands of times per second.

Since the resources might be re-used in different graphs within the
same batch (happens in tracker-miner-fs-3 for file-related content
graph data), we must be careful though not to optimize those away.
In that case we simply blow the visited resources cache on graph
changes during the processing of a batch.