core: Maintain array for refcount changes
Since refcount changes are very frequent, and each of them requires
a hashtable lookup+update to change (the refcount is the value of
the hashtable, stored with GINT_TO_POINTER), plus the replacement
goes with a new copy of the TrackerRowid for the hashtable key. Overall,
this maintenance is somewhat expensive to perform.

As there are small buffers for triple updates (64 items), a hash table
does not bring a lot of benefit in lookups, nor updates. Switching to
a simple unordered array to keep this accounting looks faster in
profiling.