core: Optimize deletion of classes of a resource
Currently the code is exceedengly thorough and issues individual deletes
for every property that is related to the class being deleted. We can take
some shortcuts here, and avoid querying existing values for most cases,
only 2 cases remain where it is necessary to fetch the previous content:

- Properties that have TRACKER_PROPERTY_TYPE_RESOURCE, in order to correctly
  change the refcounting of the resource that the property points to.
- Properties that are domain indexes in other classes, since we have to
  chain up to those tables with the right value being deleted.

All other situations can do without fetching the previous values for the
property, and for single valued properties, they can even be delegated to
the deletion of the row in the table representing the TrackerClass.

The result is a speedup while deleting entire resources.