core: Avoid finding out resource insertion through GError
When ensuring that a resource IRI is known in the database, we optimize
for newly added resources and try to insert them without further checks,
only resorting to a query to fetch the resource ID if that insertion failed.

But we figure it out after getting a GError, and clearing it. Instead, just
check the return value, so that we don't create/free errors for every resource
where that assumption fails.

Also, we failed to cache the resource if the second querying step was taken,
fix that to get another nice speed improvement.