core: Avoid query for subclasses of a resource
In the slow paths of deleting a class, we must recursively delete
all subclasses prior to the deletion of this class. We figure out
the existing subclasses of a class for a given resource through a
query, but we have this information right there.

Use the types array, and recursively handle direct subclasses of
the class being handled for deletion, so that things cascade properly
and it mostly consists of a couple of array lookups as opposed to
a database query.

This should make these slow paths faster.