core: Decouple FTS updates from previous property value lookups
Right now, FTS updates do iterate the set of previously looked up
properties in order to find the FTS ones and issue the FTS update.

We want to optimize on fetching the old values for properties, and
FTS is orthogonal to it, so decouple both machineries. As a replacement
keep a list of FTS properties that were modified, so that these are
updated when flushing.