libtracker-sparql: Fix off by one in checks for escaped IRIs
The 0x20 character should also be escaped as per the SPARQL reference,
and it correctly is when setting a TrackerResource IRI. Even though,
the fast path check for the presence of characters that should be
escaped is missing it, so it would be possible to let IRIs that only
have this invalid character as valid.

Since 0x20 (whitespace) is possibly the most ubiquitous character that
should be escaped, it's a bit of an oversight.

Fixes: 33031007c ("libtracker-sparql: Escape illegal characters in IRIREF...")