ontologies: Move fts: prefix definition to base ontology
This prefix is used in some of our base SPARQL API (e.g. fts:match
and the misc functions to deal with results), but the prefix is
currently defined in the Nepomuk ontology side.

This makes applications that want to use FTS with custom ontologies
either need to declare the prefix themselves, or help queries with
`PREFIX` syntax to add the missing stock prefix.

Move this prefix definition to the base ontology, so all databases
inherit the builtin fts: prefix and it can be used right away in
FTS queries.