libtracker-sparql: Pass cancellable to D-Bus cursor writing operations
If cancellation happens while the D-Bus endpoint is writing cursor contents
for a D-Bus TrackerSparqlConnection client, we do not much in the way to
stopping the FD writing operation from happening.

This however also means the FD was closed by the other end, so they will
fail anyway with an "Interrupted" error when writes start to fail. Make
all things depend on the same cancellable, so the error is consistently
G_IO_ERROR_CANCELLED and we bail out as soon as possible.