libtracker-sparql/core: Plug leak in triples vtab
The TrackerTriplesVTab struct created to back the tracker_triples
table up is not being properly freed. Since this is an eponymous-only
table, we never drop the table, thus the xDisconnect vfunc is the
right one for us.

In case we ever wanted to drop the table, make the xDestroy method
the same, since this virtual table creates no supporting non-virtual
tables.