core: Shuffle call to obtain subject ROWID during updates
This can be done a little bit earlier and avoid a slight bit of
busy work if the resource won't exist anyway (or errored out).
This will allow replacing the check for idempotent deletes with
a more generic approach, and fixes a (currently very unlikely)
GValue memory leak if we were to take this return path.