core: Refactor check for idempotent deletes
Idempotent deletes are those that are unnecessary to perform
because the triple subject does not even exist. This used to
apply exclusively to "DELETE WHERE { <single triple> }"
SPARQL updates.

Shuffle this check so it applies to all individual triples of any
generic DELETE operation, and closer together with the actual
database inserts.