core: Warn on non-absolute IRIs during insertion
The IRIs being inserted in the triple store should be ensured to be
correct, full IRIs. Perform the minimal check here, and warn if
the IRI does not at least contain a scheme separator.

This at least will raise some warnings if invalid IRIs like <123> or
<abc> are inserted by the Tracker user, it will be turned into a
GError worthy condition in future versions.