core: Handle blank node insertion together with all other database updates
In order to do this, blank nodes are now given an ID that is only
meaningful locally inside the TrackerSparql object, this local ID gets
translated to an actual ROWID at the time of handling inserts/deletes.
Other than that, these are still non-zero integers in the gint64 range.

Handling blank node insertion this way, the database inserts that
give the blank node a ROWID are handled in the same place than all
other database updates (namely inserting IRIs, and handling the individual
insert/delete/update triple statements).

The downside of course is having to deal with this indirection,
(a possible source of counfusion), plus accounting for a mapping between
these local IDs and blank node labels.

This indirection may feel like unnecessary complexity so far, but
it will gain some meaning when update query parsing and execution
become further separated.