core: Make TrackerToken able to store blank node labels
Currently, ROWIDs for these are generated during SPARQL update
parse tree handling. Since we want to separate SPARQL parse tree
handling from updates, and currently the mapping of blank node
labels is global (e.g. shared across a TrackerBatch, specified
from the outside at the time of creating the TrackerSparql), we
want a way to express these blank node labels so that the right
ROWID may be decided during query execution.

This is currently unused, but that will change in future commits.