core: Add code to handle post-parsing execution of updates
As we want to separate execution of the actual update commands
with the parsing of an update SPARQL string, there is now (unused)
support for handling these operations, queued into an array of
TrackerUpdateOp structs. In this list of operations, INSERT/DELETE
clauses get the (possibly complex) defined quad patterns de-structured
into a series of udpates on individual quads. This is very similar
to how they used to be handled, these operations will be queued
during parsing, instead of handled in place.

A thing to note is that an update SPARQL string may consist of
multiple commands itself, either INSERT/DELETE or graph maintenance
operations (LOAD/CREATE/CLEAR/DROP/MOVE/ADD/COPY). Besides,
INSERT/DELETE operations may be expanded into a list of operations,
and there's also DELETE{}INSERT{}WHERE{} syntax where different
sets of operations may need to be handled as a single unit. For
that there is another TrackerUpdateGroup struct array, that
delimits each of operations of the other array into groups that
correspond to the high-level query structure, and optionally
provides a WHERE clause for the places where it's needed.

These arrays must be iterated altogether in order to make a sense
of the query structure altogether. This is done in this commit,
but still unused. Future commits will port to this new infrastructure.

With this in place, we will be able to "replay" updates, without
having to parse the SPARQL from scratch.