core: Handle error differently
When raising an error if variables are used in INSERT/DELETE DATA
statements (disallowed by the spec), we indirectly rely on
solution_var_map being present (indication that this is an update
with a WHERE clause).

Make things clearer, avoid this dependency on this hashtable
(which will disappear soon), and reference the parts of the spec
that require this behavior.