bus: Extend UpdateArray FD format to accomodate TrackerSparqlStatement
Besides the SPARQL string, we need to pass a series of parameters.
Serialize those values as a a{sv} variant at the end of the SPARQL
strings.

Since there must not be \0 embedded in the SPARQL string, this
remains backwards compatible, if different versions of Tracker were
used at each side. This is however moot since we do not allow
updates to happen from within sandboxed environments.