core: Remove "_:" prefix from all blank node labels in the map
Right now, the SPARQL parsing bits add these blank node labels with
the _: prefix stripped, while the update paths from TrackerResource
do not. Make it sure that the strings in this map are consistent,
so blank node labels can be referenced from multiple elements in a
TrackerBatch, as it was intended.