remote: Implement avahi support in HTTP endpoints
HTTP endpoints may announce themselves through DNS-SD. The _sparql._tcp
service name is reserved for this kind of services:

http://dns-sd.org/ServiceTypes.html
https://www.floop.org.uk/eagle/discovering-sparql

Implement Avahi support for our HTTP endpoints, so that these are
properly discoverable.

It is worth noting that this is reserved for users of the
TrackerEndpointHttp API, and currently there are no users of this API
except the "tracker3 endpoint --http-port" CLI tool, so no data is
HTTP discoverable by default, and this change will arguably make it
more discoverable the case of any network-local HTTP endpoint being
exposed with insufficient authorization checks.