libtracker-sparql: Deprecate tracker_sparql_connection_update_blank*()
Per the spec (https://www.w3.org/TR/sparql11-query/#BGPsparqlBNodes):

    Since SPARQL treats blank node identifiers in a results format document
    as scoped to the document, they cannot be understood as identifying
    nodes in the active graph of the dataset.

This is precisely what tracker_sparql_connection_update_blank() (and async
variant) are doing. While that works given the interpretation of blank
nodes in a TrackerSparqlConnection with default flags, this is precisely
what TRACKER_SPARQL_CONNECTION_FLAGS_ANONYMOUS_BNODES implements, and at
best it is not great to advise this inconsistency as public API.

It is best to deprecate this set of functions, any user should devise
their own IRI generation scheme, instead of relying on blank nodes being
sort of one.