libtracker-sparql: Deprecate tracker_sparql_cursor_rewind()
This function cannot be honored across all TrackerSparqlCursor
implementation (direct one works, FD based ones only work as long
as the FD is seekable).

The alternative to make this call work everywhere is to buffer
the FD contents (e.g. through a GBufferedInputStream), so rewinding
happens in the in-memory buffer, but that pushes the need on all
implementation to do this and cache full documents just for the
extremely unlikely case that the cursor is rewound.

This is not great, and the operation can be easily avoided (handle
all effects from the cursor at once, or issue 2 queries), so
deprecate this API call to avoid encouragement.