libtracker-sparql: Error out if mixing TrackerSparql API between query types
Given public API misuse, It will be a possibility to try to feed a wrong
TrackerSparql into execution APIs (e.g. tracker_sparql_statement_execute() on
an update query, or tracker_sparql_statement_update() on a select query).

Add this check so that these calls error out if fed a wrong
TrackerSparqlStatement.