core: Refactor code
The handling of triples where the object is a variable is not
that much different from literal/parameter objects, only the
TrackerBinding created is different. Merge these paths together.

The variable TrackerBinding was optionally made nullable in some
paths, but the places where we didn't do so were more situations
where it could be avoided than it causing the wrong data. In
order to keep things simple, it's now set on all circumstances.