core: Add SparqlPrintValue() SQL function
This function can generically handle type conversions in a way that
we can integrate in our unrestricted and property path queries, it
takes 2 arguments for the value to convert, and the TrackerPropertyType
it expresses.

This way, we can generically handle conversions between RDF types
without being constrained to SQLite types and implicit conversions.

This function only does basic type conversion though. In order to
improve performance over SparqlPrintIRI(), this function does expect
the ROWID to IRI conversions to be done beforehand, any raw ROWID
gotten is converted to a blank node label.