core: Cluster queries in tracker_triples virtual table
Currently, the virtual table does query all properties individually,
even single-valued ones that live together with other single-valued
properties in the class table.

Since we no longer have to deal with value-to-string conversions
for the upper layers, it is now easier to query these single-valued
properties altogether in a single query to the whole class table.