core: Group tracker_triples results by graph
Graphs are currently the last component in our iteration through
tables. Move it to be on top of the loop, while it does not affect
performance querying things in one or other order, this provides
better natural clustering of data for the serialization formats
that handle graphs (e.g. Trig).