libtracker-sparql: Avoid detaching all nodes in JSON-LD serializer
If there are circular references, we might end up in a state that all
JSON nodes were made part of a subtree, leaving us with no "root"
node to unroll the data.

Make it sure that we are left at least with a single root node, that
every other nodes hang from.