core: Fallback to resource creation case if resource has no types yet
It might happen that a resource IRI is first registered, the buffer
fills up and flushes with data not yet related to this resource, and
this resource is not considered "new" anymore on later buffer handling.

Even though the worst thing that might happen here are some redundant
queries, we might still avoid them completely. So enable the "create"
flag if the resource does not have RDF types, meaning it was only
inserted to the Resource table.