core: Cache FTS delete/insert statements directly
The one piece of good news from having to do full-row FTS insert/delete
in the case of updates to FTS properties is that these queries are now
directly cacheable, instead of going through the TrackerDBInterface
LRUs.

Keep TrackerDBSstatements around for these, so that they can be
executed many times and ensured to be compiled exactly once.