libtracker-common: Add 'fts-integrity' debug flag
This debug flag forces a FTS integrity check after every set of FTS
updates, and raises an error if the integrity check did fail. This
is a more proactive (and expensive) approach to finding out FTS index
corruptions.

In order to make this helpful right away, toggle this flag on for
our own test suite, so that CI may catch any remaining/popping issues.