core: Propagate parsing errors directly on TrackerSparql creation
We used to error out on parsing errors on first execution. Make this
propagated early, so that it is possible to know at TrackerSparqlStatement
creation time whether parsing was successful or not.

This typically didn't matter since statements could only be used with
select queries, and erroring out on first tracker_sparql_cursor_next()
is alright. But now statements are used for both updates and selects,
possibly from the same entrypoint, so it's better to error out early.