docs: Migrate documentation to gi-docgen
Move away from hotdoc, and port the documentation infrastructure
to use gi-docgen. We still do fixup devhelp2 files in order to
include ontology items as search terms, and we still do prefer
to ship the pre-generated documentation in the tarball, in order
to avoid a host of dependencies like graphviz for the people
building from tarballs.

It is not the first time that porting to gi-docgen was attempted,
this is partially based on work by Daniele Nicolodi <daniele@grinta.net>.