libtracker-sparql: Use setlocale() directly to query locale
Long long ago, this was an integration point since the Maemo platform
(I thought I'd never write that in 2023) had some infrastructure to
allow on-the-fly locale changes.

This is long long gone, and nowadays it's just a setlocale() wrapper
with a recursive lock that does nothing since all MT-unsafety of this
function comes from the global environment.

We can just setlocale() directly to query the locale details, with
pretty much all the same guarantees.