manpages: Do not recommend broken API commands
The flatpak CLI parsing for additional policies changes at some point
so that it is not possible to specify multiple values in a single call.
Doing that results in the ';' being escaped, and the full string to be
interpreted as a single value.

In order to add multiple values, --add-policy can be called multiple
times, recommend that in our manpage for the task.