core: Handle FTS search terms individually
Commit da9eb9a07c hid the internal FTS syntax as an implementation
detail, but it did so by making the full search string a single term,
so multiple input words had to also appear consecutively in the searched
text to be considered a match.

Bring back the flexibility in order/distance of the looked up words
by using a SQLite helper function to split FTS search terms. This
still results in the FTS syntax being hidden away, but splits the
input words so that each becomes its own search term.

Fixes: da9eb9a07c ("libtracker-data: Do not expose FTS5 syntax through fts:match")