ci: Run tests for libsoup 2.x
Since we are departing from test beds that have libsoup2 only
into ones that have soup2 available but soup3 is also available and
enabled by default, the several paths skewed the coverage testing
paths.

Add a test target for soup2 specifically, so we get some of it back
and spot regressions in libsoup2 support.