libtracker-sparql: Improve threading guarantees of TrackerNotifier
Make the TrackerNotifier::events signal ensured to be emitted in the
same main context that created the TrackerNotifier. This is coherent
with other threaded objects that have communication through signals
(e.g. GDBusProxy).

Notably, it is already expected to behave this way in tracker-miners.
Since the tracker-miner-fs-3 D-Bus endpoint is managed on a distinct
context/thread that should also handle the proxying of notifier events
without involvement of the main thread/context.