libtracker-sparql: Check cancellable more often in TrackerNotifier
While handling the URN query cursor, we did not check the cancellable
and went all through with it, even if TrackerNotifier finalization
paths might have cancelled the task. This could not happen in practice
since the cache struct takes a reference on the TrackerNotifier, but
will happen with future commits.