libtracker-common: Initialize stemmer once
We currently initialize the stemmer on TrackerLanguage init(), then
after the language is actually set. Since that now happens just a
little bit later on constructed(), we can avoid the double initialization.

Also the locking around stemmer is superfluous, since the TrackerLanguage
object was not used nor given anywhere yet.