libtracker-common: Get default language from environment
Currently, passing a NULL language will invariably result in
english being used by default for stop words an stemming. And
unfortunately, we do pass NULL expecting the TrackerLanguage to
find the better language for us.

Make it actually do so, and pick the default language from the
environment, and only add english stop words in addition if
English is specified in LANGUAGES besides the current locale.