remote: Fix possible double free
In case of error reading from the input stream, we would
propagate the error through the GTask and break, falling
through paths that also try to return and free the task.

Handle the task entirely at the end of the function, based
on the return value. We still need to close the input stream
and soup message in any case.

CID: #1517496