tests: Adapt FTS snippet tests to run before/after SQLite 3.42.0
This SQLite version fixes a small bug in the FTS5 snippet function
that affects our output checks here. Concretely, it might ignore
the specified number of output words with repeated matched terms.

It is good that we have FTS tests, but we do not need to test for
this behavior specifically. Adapt the tests so that we keep their
spirit but bypass the SQLite behavior change. The snippet-4 test
coincided too much on it and had to be dropped, it's a minor loss
though.

This fixes the test suite with SQLite >= 3.42.0.

Closes: https://gitlab.gnome.org/GNOME/tracker/-/issues/405