build: Avoid meson.build_root()
This gives the "wrong" answer when the project is embedded as
subproject in a larger one, as it'll return the root of the main
project rather than the root of the Tracker SPARQL source tree.

Use a variable set by calling meson.current_build_dir()
in our toplevel meson.build instead.