Revert "build: Avoid meson.build_root()"
This was a small oversight. The expected value here is the root
project build root, since Tracker module loading bits want to
detect the case that it running uninstalled in-tree, and it
does so by checking that the current dir is the build dir.

When building as a subproject, this subproject root does not match
with the global build dir, so the detection of this situation
is fooled into looking at the installed modules.

Reverting this will fix test failures in tracker-miners repo like:

  Tracker:ERROR:../subprojects/tracker/src/libtracker-common/tracker-parser.c:115:ensure_init_parser: assertion failed: (module != NULL)
  (test program exited with status code -6)

While at it, add a comment explaining why we so far need the
root build dir.

This reverts commit aa2ab3855a6a381b2e4e420ab3f9cb4cce8a24a2.