core: Pass TrackerOntologies when creating FTS tables
An unmaterialized benefit of the library unification done in the past
is that we can pass the TrackerOntologies object to the code in charge
of creating the FTS table, instead of preparing data for consumption.
This simplifies code, and allows looking up for things that we previously
couldn't.

As a side effect, column ordering in the FTS table now changed, which
affects fts:offsets and fts:snippet output, the tests had to be slightly
adapted to that.