utils: Fix trackertestutils .pc file pythonpath as expected by miners
Commit 45e070a5dc fixed python_path in trackertestutils .pc file to
have a relative path, but also appended a '/trackertestutils' for easier
consumption for the "command" variable definition.

This "python_path" variable is used from external projects (e.g.
tracker-miners) to find out the trackertestutils python module location,
and should keep pointing to the same directory level. This results
in errors like:

  ModuleNotFoundError: No module named 'trackertestutils'

When trying to run tracker-miners test suite with tracker installed
out-of-tree.

Shuffle this piece of the path from the "python_path" variable to
the "command" one, so that everything uses relative paths but
python_path keeps pointing to the expected location.