libtracker-sparql: Change binary format of DBus updates FD
We've drawn ourselves to a few corners wrt extensibility of the
FD binary format, in order to accomodate further TrackerBatch
operations.

ABI is not a concern here, while the DBus API stands a bit on
the grey area, there are in practical effects no implementors
of the DBus API to constraint ourselves to.

Also, compatibility is not a concern, updates are to begin with
forbidden by the portal, so there's no possibility of mismatching
versions here. Updates on DBus are pretty much between system
components like miner/extract, hence they ought to run the same
tracker library versions.

The changes in the format are namely:
- SPARQL operations have now a "header" with the TRACKER_BUS_OP_SPARQL
  opeartion type.
- SPARQL string and the possible parameters afterwards are dealt with
  as separate strings, instead of the "parameters stance behind the \0"
  trick.