core: Set FTS weight on TrackerProperty objects
Before, we didn't look for this value in TrackerProperty, that changed
with the switch to BM25 where we use TrackerProperty to set up correctly
the FTS table and its ranking function. However we mistakenly assigned
the same default weight (=1) to all properties.

Initialize this value when reading the ontology files, so the FTS rank
function gets the correct weights during the necessary changes.

This again (sigh) changed expected outputs of some of the ranking tests,
so adapt to that.