core: Re-invert BM25 ranking order
As per https://www.sqlite.org/fts5.html#the_bm25_function:
"The "-1" term at the start of the formula is not found in most
 implementations of the BM25 algorithm. Without it, a better match
 is assigned a numerically higher BM25 score. Since the default
 sorting order is "ascending", this means that appending
 "ORDER BY bm25(fts)" to a query would cause results to be
 returned in order from worst to best. The "DESC" keyword would
 be required in order to return the best matches first. In order
 to avoid this pitfall, the FTS5 implementation of BM25 multiplies
 the result by -1 before returning it, ensuring that better matches
 are assigned numerically lower scores."

This behavior is basically opposite to the previous default ranking
function, and Tracker's own, where we did use to give higher values
to better matches and expect DESC order. Re-invert this behavior in
queries, so we don't break compatibility and keep assigning higher
values to better matches.

Since this was quite an oversight (shadowed by the algorithm change),
the FTS ranking tests had to adapt to the new current ordering.