docs: Redirect to website, take 2
The initial attempt failed, as seen here:
<https://gnome.pages.gitlab.gnome.org/tracker/_redirects>

The error was:

    rule 1: error: no domain-level redirects to outside sites

To get around this, add a HTML page that redirects to the new website.
Requests for /faq will now go to the main page, but it shouldn't be too
hard to find the FAQ link from there.

Overview links are removed as we already hav
<https://gnome.pages.gitlab.gnome.org/tracker/overview.html>.