website: Use 301 (redirect) instead of 200 (rewrite) for ontology redirects
As nice as it seemed to preserve the same URL in the browser location bar,
this does not play nicely with gi-docgen looking up assets relative to the
location (and failing). This results in CSS-less HTML being served.

Since it's not clear that gi-docgen will want to consider the usecase,
use 301 redirects, which will change the location bar but will make the
website look as expected.