docs: Add CamelCase/dromedaryCase anchor targets on ontology docs
These will come in handy to convert our own ontology links (e.g.
".../nmm#MusicPiece" to a link in our documentation. The stock
behavior of gi-docgen is to lowercase document anchors, so it
does not help for our case where capitalization is important.

Adding the anchors in the generated markdown ourselves will
propagate to the generated HTML, thus help with our redirects.