website: Fix redirection links
Now that we 301 the ontology links, it turns out the target
is not what Gitlab expects, resulting in 404 as an extraneous
"tracker/" folder is added after redirection. from
https://docs.gitlab.com/ee/user/project/pages/redirects.html:

If your GitLab Pages site uses the default domain name
(such as namespace.gitlab.io/projectname) you must prefix every
rule with the project name [...] If your GitLab Pages site uses
custom domains, no project name prefix is needed.

With the tracker.api.gnome.org subdomain involved, we are in
the second case, and that is where we are interested in these
redirects working. Change the redirect targets so it works as
expected in that domain.