core: Handle SelectClause between group/having and order/limits
This is the right order as per the SPARQL spec
(https://www.w3.org/TR/sparql11-query/#convertGroupAggSelectExpressions)
in order to get in-scope variables right at each step. We used to have
some degree of leniency here with undeclared variables due to "" vs. ''
handling in SQLite, but it is better not to rely on that.

If there is no solution modifier, the SelectClause node is still handled
in place.