direct: Unify select/update query data in a single struct
We used to have a wrapper TaskData, with task-specific data/destroynotify,
since we want additional payload on every task, and the
GTask->TaskData->per-task-data indirection gets a bit ridiculous, make
TaskData contain a union of all the structs so it gets allocated and dealt
with as a single block.

Since there are more task types coming, this will also ease their handling.